# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ArticleArchiveView": {
        "properties": {
          "error_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Class"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "source_url",
          "status"
        ],
        "title": "ArticleArchiveView",
        "type": "object"
      },
...
      "BrowserCaptureBody": {
        "properties": {
          "html": {
            "minLength": 1,
            "title": "Html",
            "type": "string"
          },
          "source_url": {
            "minLength": 8,
            "title": "Source Url",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "token": {
            "minLength": 8,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "source_url",
          "html"
        ],
        "title": "BrowserCaptureBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Response_ArticleArchiveView_": {
        "properties": {
          "code": {
            "default": 200,
            "description": "业务状态码:成功 200,失败非 200",
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArticleArchiveView"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          },
          "message": {
            "default": "success",
            "description": "提示信息",
            "title": "Message",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应时间;默认不携带,不携带时省略",
            "title": "Timestamp"
          }
        },
        "title": "Response[ArticleArchiveView]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Verso",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/me/articles/browser-capture": {
      "post": {
        "operationId": "capture_from_browser_me_articles_browser_capture_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowserCaptureBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_ArticleArchiveView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capture From Browser",
        "tags": [
          "article"
        ]
      }
    },
...
  }
}